Restructuring standard specs
