+++ spec/common_scripting_layer/bar/bar_gen.rb
