+++ spec/common_scripting_layer/bar/bar_spec.rb
