+++ spec/common_scripting_layer/bezier/bezier_spec.rb
