+++ spec/common_scripting_layer/bubble/bubble_spec.rb
