+++ spec/common_scripting_layer/dot/dot_spec.rb
