+++ spec/common_scripting_layer/error_bars/error_bars_spec.rb
