+++ spec/common_scripting_layer/histogram/histogram_spec.rb
