+++ spec/common_scripting_layer/line/line_spec.rb
