+++ spec/common_scripting_layer/scatter/scatter_gen.rb
