+++ spec/common_scripting_layer/scatter/scatter_spec.rb
