+++ spec/common_scripting_layer/spec_helper.rb
@@ -1,9 +1,9 @@
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'rubyplot'
-require 'bar_gen'
-require 'dot_gen'
-require 'line_gen'
-require 'scatter_gen'
+require 'bar/bar_gen'
+require 'dot/dot_gen'
+require 'line/line_gen'
+require 'scatter/scatter_gen'
 
 def compare_with_reference?(test_image, reference_image, tolerance)
   compute_rms(test_image, reference_image) < tolerance
