+++ spec/common_scripting_layer/stack/stack_spec.rb
