+++ spec/common_scripting_layer/stacked_bar/stacked_bar_spec.rb
